<feature> Generation plans (#810)
Add support for an extra subset called "genplan" which is run before any
other template processing.

It produces a small script that is then used to control the passes and
alternatives used during template generation. If no plan is produced,
then the current defaults apply.

Implementation of generation plans will result in less invocations that
generate empty results and speed up overall processing time.

There are a couple of examples in the commit for sqs and rds. Note that
if a plan is produced, component processing can immediately be
terminated to make plan generation passes as quick as possible.

As rds is the only component that needs more than one alternative, the
alternatives default has been wound back to just have the primary for
solution templates.

A second deployment plan subset has also been allowed for. It will
permit overriding of the default deployment behaviour for a component.
If required, a genplan should be added which includes the depplan. An
example where a deployment plan is required is for certificates, where
the deployment process needs to trigger but not wait for the certificate
template to complete, in order to obtain the necessary DNS entry
information to permit validation of the domains.