<feature> Generation plans for all components
Add explicit generation plans for all units, and simplify the defaults
accordingly.

Rearrange the decision tree used to determine the output format required so
that debug does not interfere with the generation of plans.

Fix a couple of variable references for RDS creation. The use of local
only variables in component templates highlighted a couple of incorrect
variables references that relied on global variables.