<feature> Resource Groups and placements (#827)
* <feature> Resource Groups and placements

Allow different subsets of a component's resources to be deployed
to different accounts/regions.

The subsets are called Resource Groups. The set of attributes defining
the deployment details is called a placement.

For now, placements are assumed to provide provider and deployment
framework information. A subsequent change will look up the provider
based on the account. A default provider of AWS and deployment framework
of cloud formation are used for backwards compatability.

The component designer decides which resources can be deployed to different
placements, thus deciding how many resource groups are needed.

Each group is then given a name in the component configuration, independent
of the provider. Each group is also assigned a name for the
placement to be used.

Every component has a "default" resource group where the bulk of its
resources are typically created. It uses "PRODUCT_PLACEMENT" as its
placement name.

Each resource group can have "shared" attributes which are expected for
all implementations. It can also have provider specific attributes which
will be expected depending on the placement of the resource group.

Placement names are mapped to actual placement details via a placement
profile. Placement profiles can be configured at the product and tenant
level, and support qualification.

When a component needs to be processed, the logic first needs to
determine the placement details to determine the provider to be used.
This in turn defines the set of attributes required, and the processing
needed to create the component and manage its state.

To implement the above, the macros for creating the component
configuration have been adjusted to create the default resource group
for each component.

The template loading logic was extended to work off the resource group
placement information, along with the process of determining the
processing macros to be used.

The handling of special attributes used for profile and placement
handling, along with the standard attributes all component have, was
modified such that they are added to the configuration provided by the
component configuration files. This permits documentation generated from
the component configuration to show the full range of configuration
options. Attributes are ordered so these framework injected attributes
appear last in the list of attributes, so that documentation can focus
the reader's attention on the unique attributes of the component.

A hook has also been added to permit each provider specific
implementation of a component to specify the provider specific services
it uses, and for these to then be loaded whenever that implementation is
required. A subsequent change will use this mechanism to move all of the
remaining service configuration in the original template tree across to
the new provider based tree.

* Use resource group as placement profile index

Standardise resource group keys and use them as the index into the
placement profile, rather than an explicit Placement attribute of the
resource group.

The combination of profiles, qualifiers and PlacementProfiles still
provides enough flexibility to handle currently foreseen scenarios.