<refactor> Template processing
This is the first of several PRs to rework the processing of templates
to add multi-provier support.

The first step is to factor out some generic processing for standard
output formats (text, json), which is what this change does.

Next step will be to make the top level macro for a template
dynamically.