<refactor> Convert Exception to Fatal
Align to logging terminology so a previous "exception" is now a "fatal"
error.

Replace @cfException with the more general @fatal. This also makes the
parameters for any logging consistent, with description= replaced with
message=.

Add the ability for any logging to provide a message, context and detail
in order to make the macro conversion straight forward.