<refactor> tidyup aws use of dynamic outputs (#839)
* <refactor> Remove cf specific macros

Replace cfScript, cfDebug, cfCli and cfConfig with the equivalents using
the generic output format management macros.

Add a macro wrapper around getGenertionPlan to simplify the syntax for
adding a genplan to a component.

* <refactor> Convert Exception to Fatal

Align to logging terminology so a previous "exception" is now a "fatal"
error.

Replace @cfException with the more general @fatal. This also makes the
parameters for any logging consistent, with description= replaced with
message=.

Add the ability for any logging to provide a message, context and detail
in order to make the macro conversion straight forward.

* <refactor> Introduce pre/post condition macros

Replace existing cfPreconditionFailed and cfPostconditionFailed with
their generic equivalents.

* <refactor> Remove mode

With output processing now single pass, mode is no longer needed.

* <refactor> Remove last of mode

Remove mode as a parameter on all macros.