<fix> Application level templates
Correct the application template to include the correct level.

Treat the absence of any suitable processing macro for a unit as a fatal
error.

Correct the detection of fatal errors in scripts - the length of the
severity indicator was increased to allow for "timing" but the check for
fata was not adjusted accordingly.