<feature> First cut of context based in-memory model
In order to accommodate accounts and regions as well as cross-product
links, introduce a new in-memory model for processing based on contexts.
It has been designed to make link processing efficient along with
calculating hierarchical overrides and qualification.

So far the code generates a simplified model but doesn't do anything with it.

In order to assist in debugging the new model, a special unit of "model"
has been added to the application level. It produces a config file
containing the model prior to processing commencing.

This commit also fixes a bug with link processing where
instance/versioning processing was wasn't correct.

It also fixes a couple of instances where calls the standardPolices
weren't including the now mandatory baseline links parameter.