<refactor> Move master data under respective provider (#844)
* <refactor> Move master data under respective provider

Provider inclusion process now loads the master data. For now its
wrapped in an ftl file until we can directly pull in the json.

Note that Id and Name atributes need to be explciitly provided where
expected by code, as the files are not not subject to the Id/Name
injection that occurs for the blueprint.

* <fix> Incorporate master data changes from other commits

Add alerting master data into the shared master data file.