<feature> Dynamic model control.
Introduce support for dynamic functions as well as macros.

The calling syntax means parameters are provided as an array but that
seems reasonable is a known set of parameters are always provided, even
if empty.

Also introduce the concept of the "model". This is the internal
representation of the CMDB expected by a deployment framework. Support a
legacy format to cover the current processing, and introduce a new
default which leverages the context tree approach. The model uses a
dynamic function to determine its contents so follows the same sort of
pattern as output processing.

Flesh out the context support to allow creation of a context tree
equivalent to the currently loaded blueprint.

There's more work to do in regards to occurrences but this commit
avoids a growing mega-commit for multi-provider support.