<fix> Ignore iam/lg resources for deployment state (#877)
Iam and lg resources can be created via the iam and lg units before the
corresponding occurrences actually exist.

This commit adds a flag that can be put on any resource in the state of
an occurrence. The flag controls whether the resource is considered when
deciding if the occurrence is deployed or not.