<refactor> Use occurrence to determine baseline links
Refactor the getBaselineLinks routine to accept an occurrence and use
this to determine the baseline profile.

The next step as part of the move to use of the context based model will
be to use the occurrence in the call to getLinkTargets, so that the
correct baseline is used in the situation of multiple products being
available simultaneously.