<feature> Round out context based model support
Support full template generation via the "context" model.

It is now backwards compatible with existing subcomponent use, but
depending on how the setup routines are defined, can also allow
subcomponents to be created independent of their parents. This should
enable a smooth transition of services and tasks to be independent
components.

Text output generation has been refactored to support the ability to
have multiple sections in the output, define the order in which the
sections are serialised and allow for a "converter" per section to
convert the data structure used in the section to a text format. This
feature is used to ensure that generation passes are only run once, even
though requested by multiple components via the selected deployment
unit.

Fix a bug where an "Enabled" setting was being added to te environment.

Also add explicit id/names for some internal links.