<feature> Autoscaling ECS & Ec2  (#952)
Uses AWS Application AutoScaling ( for some reason AWS has 3 autoscaling services ) to configure autoscaling for ECS based Services

Autoscaling basically uses cloudwatch Metrics/Alarms to update the desired count of services. So we use our existing metric and alarm setup combined with links to enable a service to track the metrics of another component. Services can also track their own metrics if a link is not specified

Autoscaling is configured using ScalingPolicies on a service. Multiple policies can be applied to a service and the one which results in the highest desired count wins.

Scaling Policies have 3 different types

Stepped - An alarm is created which sets a water mark level - Adjustments are then made relative to the water mark - Changes the desired Count of the service

Tracked - You specific a metric and a target value and AWS will automatically scale the services to maintain the metric at the target value - Changes the desired Count of the Service

Scheduled - Allows you to change the Min and Max counts of the service to limit the scalability of the service based on time ( e.g. business hours vs after hours )