<refactor> consolidate command line options (#992)
Create a single variable that manages command line options. It is then
clear in the code when a command line option is being used.

First use case is the introduction of a command line option o control
the resouce group processed, which as a global variable would clash with
localised uses of the same variable name.

Conversion of all use of command line options to the new structure can
happen incrementally.