<refactor> Finalise use of commandLineOptions
Update all uses of global variables representing command line options to
use the single commandLineOptions global variable instead.

The ability to provide the build reference and build unit on the command
line has been superceded by the accessing of this information via
settings so remove the option for these values to be provided during the
template creation process.

Also remove the "Multi" template, as part of deprecating the use of
levels.