<feature> StackOutput dynamic loading and input sources (#1007)
Updates the StackOutput functionality to use some of the newer dynamic loading functionality

It introduces the idea of input sources as a new data collection, input sources should be attributes of literal objects or settings that should be included in template processing. ( Stack outputs, settings etc)

The input source types are fixed at the provider level, since they should be the same across providers to ensure compatibility and enforce the idea of multi provider lookups

However providers can override how the data is loaded into the input source. This allows us to have mock data providers or to format data during the loading process rather than formatting it outside of the freemarker engine.

Providers can also add optional default filters to the output lookup process, this functionality was added primarily to make getStackOutpuObject generic.

The default input source is known as composite which uses the composite stack outputs provided via the command line inputs.

A new import source mock has been added which allows for outputs to be mocked for template generation processing. The intention of this is to introduce a unit testing framework with mocked outputs