<feature> testplan generation (#1030)
Adds support for the generation of testplans based on testcases

This allows for external testing platforms to receive a set of assertions and test cases and to use those assertions against the template generated using one or more scenarios defined in the test case.

The deployment unit is defined in the testcase to allow for scenarios with multiple deployment units defined due to dependencies but still defining which template output you want to test

Command line options are now their own input type which allows for any command line option to be overridden for both mocking and test cases.

The testplan is a new output type which is generated if a testcase has been specified. It is automatically added to the generation plan since we can't add it at the component level ( testcases are blueprint level not component )

The composite validation functionality has been moved up into common from component so that it can be used across other input sources

Removed the product level templates completely as they are no longer supported

Updated setContext.sh so that it only assembles the composites if the GENERATION_INPUT_SOURCE is composite - This is part way to making createTemplate.sh usable without actually having a CMDB

masterData is now treated like the other inputsources and can be loaded for each inputsource. A shared inputsource is also available so that the masterData for a provider is available when all other input sources are used. As a result it has now moved in to the inputsources directory