<Feature> Component type wildcard support in profiles (#1099)
In a same way that a deployment profile can be made to apply to all
modes via the use of a mode of "*", add the ability for a deployment
profile to apply to all component types by the user of "*". This allows
configuration applying to all components sharing a deployment profile to
be separate from that specific to the component type.