<Feature> Add external tier for external services
Need to tweak the external link processing to check there isn't
actually an external tier.

If an external link is really required even if an external tier is
present, a link type should be added.

Also a couple of typo fixes for setting namespace management as well as
a correction to the API gateway logic to manage AWS regions in URLs.