<refactor> v2.0.0 restructure of CMDB (#1107)
Reorganise cmdb to make it easier to manage via branches and dynamic cmdbs

State is now in its own directory at the same level as config and infrastructure
Solutions is now under infrastructure
Builds are separated from settings and are now under infrastructure
Operations are now in their own directory at same level as config and
infrastructure. For consistency with config, a settings subdirectory has been
added.

With this layout,
- infrastructure should be the same across environments assuming no builds are being promoted
- product and operations settings are managed consistently
- all the state info is cleanly separated (so potentially in its own repo)

/config/settings
/operations/settings
/infrastructure/solutions
/infrastructure/builds
/state/cf
/state/cot

If config and infrastructure are not in the one repo, then the upgrade must
be performed manually and the cmdb version manually updated.

Functions for context tree access now support the pre and post v2.0.0
cmdb directory formats.

Previously defined global variables have been redefined to permit
support for pre and post v2.0.0 cmdbs.

NOTE: v2.0.0 has not been included in the update list, so switching to
v2.0.0 will be done via a separate PR. For now, the aim is to test the
changes do not cause issues for pre v2.0.0 cmdbs. Manual testing of
v2.0.0 upgrades can also be performed.