<feature> Control link attributes included in context
Add the "IncludeInContext" link attribute to permit only a defined set of
attributes to be included in the context as a result of a link. This
permits environment space to be conserved where only one or two
attributes are needed. The default behaviour is all attributes if the
attribute is missing or empty.

Also change the order in which getFinalEnvironment assembles variables
to permit link attributes to be overridden by component attributes (e.g.
link points to an external service with a default timeout value that the
code wants to override). Anything explicitly set via a fragment file
still overrides everything.