<feature> Optimise cmdb version checks (#1113)
Mostly there are no cmdb upgrades to be done. On this basis, check the
last version to be checked first and only check the full set of versions
if an upgrade to the last version is required. This assumes the version
lists are ordered, which they are to ensure an upgrade can assume the
cmdb is in the state expected as a result of the previous version
upgrade.

Also fix a minor syntax error with local vs global variables in a bash
script.