<refactor> Stack Output Processing into Engine (#1108)
moves stackoutput processing from the execution ( jq ) into the engine. Currently the execution stackoutput processing assumes that the output data is aws cloudformation data and converts to our StackOutput Syntax based on that

This PR creates the composite stackoutput file as an array of objects with two properties

{ 
    "FileName" : ""
    "Contents" : []
}
This is a generic format that handles any json data

Then during the bootstrap process each provider is responsible for parsing these files based on their expected output and adding the stackoutputs it matches on

Moved assembly of settings, stackoutputs and definitions to only run during the createTemplate script. With the inclusion of build blueprints and their extended usage these composite objects should only be used by the template creation process. They shouldn't be available to other plugins ( being run* or manage* commands ) if they need these values they should access them through create* scripts which allow for us to maintain backwards compatibility and lookups in the engine

Reworked manageCrypto which looked up cmk arn values based on the old stackoutput format. Instead when a key needs to be found for both account and segment level the key is found using the createBuildblurprint engine view which provides the details as attributes of the baseline cmk keys

To include the account level cmk an extra key engine type has been added to the masterdata and to the component cmk-account which does the ID to ARN translation required to lookup the key

replaced the encrypt_file routine in setContext with a new one in utiltly which is now used to encrypt the ssh baseline key based on the key alias rather than ARN