<refactor> Move composite processing into freemarker
In preparation for direct file querying from freemarker, move much of
the logic to differentiate between cmdb files into freemarker.

The bash scripts now mainly ignore "dot" files and leave the processing
based on filenames to freemarker.

While removing a number of directory scans, most of the time in assembly
is spent in running jq on each file, which hasn't altered with the
commit. However, by moving the actual processing into freemarker, we are
ready to bypass the file scanning in bash and do it directly from the
freemarker wrapper.