<fix> Lambda policy dependencies if role in same template
If the policies are created in the same stack as the lambda function,
then they must be included as dependencies of the lambda function.

This is because they reference the role they refer to, but don't
directly link to the lambda function. This can cause race conditions
where the role is checked for permissions before the policies are
complete.

This situation arose for a lambda function being used as an
SQS event source, where the source object checks that the lambda
function has permission to read the queue before creating the event
trigger. In this case, it just depends on the function existing, which
in turn was only depending on the role being present.