<feature> X-Ray control
Add initial support for AWS X-Ray based on the components that support it via
Cloudformation - currentlt API Gateway and Lambda.

Others will be added as Cloudformtion support becomes available. The
default config for other services is passthrough which should be
sufficient given most flows start with either an API Gateway or KLambd
call.