<refactor> OpenAPI support
Revise the handling of swagger/OpenAPI specification as follows;

- use the term "openapi" rather than "swagger" in variable names etc
- add additional support for v2 vs v3 differences
- add support for lambda authorisers
- add support for OAuth2 and scope based security
- use getCompositeObject() to validate the extension config
- change the configuration of security to align more with the strucure
used by the openAPI specification
- support the configuraiton of scopes as part of the base definition
- deprecate the special global settings for signv4, api_keys and user
pools in favour of a "security" element
- add a global option to inhibit the automatic adding of "options" verbs
if not present in the definition file.

Also add a fatal error if a subobject does not have object content.
Previously the subobject was simply ignored making it hard to debug the
issue.