<feature> Dynamic CMDB loading (#1159)
Using the new getCMDBTree() function offered by the freemarker wrapper,
support the dynamic loading of CMDB information by the freemarker
enginerather than having to preassemble this information prior to
execution of the freemarker engine.

This PR introduces engine support functions for CMDB access, along with
the necessary extra switches on the freemarker wrapper to support
getCMDBTree() but does not change any processing logic to use the new
support functions. These modifications will be the subject of subsequent
changes.