<fix> cmk to baseline transition
The legacy oai credential was not being removed in the transition from
cmk to baseline.

The checks to delete the legacy oai/ssh pseudo stacks were also not
compatible with cmdbs >= 2.0.1 as they assumed the baseline and cmd
stacks were in the same directory. The CF_DIR is now adjusted
accordingly in addition to the filename.