<feature> Plugin Introspection
Using the getPluginTree() function introduced in the new version
of the wrapper, determine at runtime what providers are present
in the plugins provided on the command line.

The core files for a provider are automatically included when it is
included, but services and components are left to be lazy loaded.

As the process knows which templates are present, it directly includes
them rather than using the includeTemplates macro.

Rename the provider specific top level ftl file to become a provider
"marker" file, along the lines of the various marker files used in
cmdbs.

Tidy up the bootstrap a bit my moving the logic of numeric/text log
levels into logging.ftl and the seeding of input data into a seed.ftl
file under engine/inputdata.