<feature>Jenkinsfile for template testing (#1233)
Add an initial testing process to validate template generation and content.

The initial script focuses on running the aws plugin templates using the awstest provider. The process uses the unitlist to generate the list of deploymentunits in the blueprint and then runs createTemplate.sh on each level using the same deployment unit. This is the only way at the moment to know which unit is required.

We don't worry about template generation failing, instead we are just focussing on running any test cases which have been generated