<feature> Registry scopes (#1235)
Introduce the idea of a registry having a "scope", with "account" level
being the current default.

The initial use case is to be able to reuse the same deployment unit
value in different segments of a product, and to associate each use
with its own image, rather than sharing images as is currently
supported. This change thus introduces "segment" scoped registries.

The scope value is obtained via the "Scope" attribute of a build.

This change needs to be introduced in parallel to or ahead of the
equivalent automation change which will see images starting to be segment
scoped when stored.

This change is an interim approach until code units are introduced,
with the details of the registry to access being attributes of the code
unit that can then be queried via buildblueprint.