<fix> Build blueprint production
Build blueprint generation was failing before it ws forcing the
extension of its genplan to be .json.

Tweak the output file expected by all the consumers so the standard
output suffix as included in the uild genplan will work correctly.

Also fix the generation of a blueprint to only force the utput suffix in
the pass that genertes the blueprint.

The investigation was triggered by an attempt to encrypt credentials
using manageCrypto.sh - one of the clients for createBuildblueprint.

This change thus includes some extra debug messages and ciphertext
prefix handling, along with a fix to the handling on in file attribute
updates.