<fix> Segment checks
Only check the current segment value matches the blueprint if a current
value for product has been defined.

The segment is defaulted to "default" in the automation setup processing,
so we don't want to trigger the segment check unless we are processing a
product (as opposed to an account for instance).