<fix> Legacy state directory structure detection
The check for existing files in the root of the state directory was
including directories as well as files. This resulted in it treating
units as legacy for whom there was already a unit that contained the
name of the new unit.

An example is a pre-existing directory of bff-v1-imp and a new unit of
bff-v1.