<refactor> Updated freemarker wrapper
Update to v1.11 which separates the inialisation from the searching of
CMDB and Plugin trees.

It also added the ability to provide a hint as to the files the scanning
process should expect which is used to control the file glob pattern
used. It needs to align to any regex used but where a specific file is
expected, it should minimise the amount of files that need to have the
regex applied.