OAI cleanup fixes
Keep the logic to check if the legacy OAI needs to be removed until
there is no use of the OAI.

Use the --query switch on the aws cli to simplify the code if an OAI is
being used.

Add an extra attribute to the environment configuration definition to
permit flow log expiration to be set at the environment level. This
attribute was not included when validation of environment configuration
was added.