<feature> Contract output generation support (#1240)
Introduce the idea of contracts outlined in #1231 based on the schema defined in #1232

The idea is that the engine generates contracts which outline a set of tasks which need to be completed in order to perform a piece of work.

Tasks can be scheduled and ordered through two levels, stages and steps. Stages group steps and specify how its steps can be executed, either parallel or serial at this stage.

The first implementation of this which is part of this PR is the Generation Contract. This replaces the existing genplan which essentially outlines the same thing but without the execution ordering.

The Generation plan creates a stage per template alternative along with an extra stage for the pregeneration subset. Each of these stages as a task process_template_pass which runs the freemarker engine.

The standard generation stages are configured for parallel execution with pregeneration run in serial. This is to allow for other tasks to be run during pregeneration which depend on the initial script.

Tasks

Tasks are documented in the engine, with the expectation that the driver making the call can support each tash. Each task has a set of parameters that it accepts and they are validated in the same way that we validate our other input sources

Bash driver

This PR adds initial support for parsing contracts, the initial implementation reads through the stages and steps in the contract and generates a single line function call with the type of the task and the parameter values as a delimited list next to it

E.g. for the step
```
          "Id": "generationXprimaryXconfig",
          "Type": "process_template_pass",
          "Parameters": {
            "provider": "aws",
            "framework": "cf",
            "outputType": "json",
            "outputFormat": "",
            "outputSuffix": "config.json",
            "subset": "config",
            "alternative": "primary"
          }
        }
```
a line in the contract command file will be created as

core_process_template_pass aws;cf;json;;config.json;config;primary
The delimiter can be configured through the utilty task. The intention with this setup is that if the delimiter " " is used you could potentially execute the created script directly to run the tasks provided. We don't do this as of yet.